Refactor auth tests to separate module
